reorganize and cleanup r3f components
